+++ EnjoyTripBackend/src/main/java/com/example/EnjoyTripBackend/util/DataUpdater.java
@@ -1,5 +1,6 @@
 package com.example.EnjoyTripBackend.util;
 
+import com.example.EnjoyTripBackend.util.collector.AirplaneInfoCollector;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
