+++ EnjoyTripBackend/src/main/java/com/example/EnjoyTripBackend/util/argumentresolver/SessionConst.java
@@ -1,4 +1,4 @@
-package com.example.EnjoyTripBackend.util;
+package com.example.EnjoyTripBackend.util.argumentresolver;
 
 public class SessionConst {
     public static final String LOGIN_MEMBER = "loginMember";
